refactor: 💡 use company repo instead of account book repo
✅ Closes: #4484
